Move ProcessingBuffer into separate file
